% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{coefficient of variation}
\usage{
cv(x)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
cv
}
\description{
Compute the coefficient of variation
}
\examples{
set.seed(12345)
x<-rnorm(25,2,3)
cv(x)

}
\keyword{cv}
