% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errortipoI.R
\name{errortipoI}
\alias{errortipoI}
\title{errortipoI}
\usage{
errortipoI()
}
\value{
errortipoI
}
\description{
del error tipo I en Estad
}
\examples{
errortipoI()
}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
}
\keyword{errortipoI}
