% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_mean.R
\name{g_mean}
\alias{g_mean}
\title{Geometric mean}
\usage{
g_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}

\item{na.rm}{Should missing values be removed.}
}
\value{
a geometric mean
}
\description{
get a geometric mean
}
\examples{
x<-rnorm(25,2,3)
g_mean(x)

}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
}
\keyword{g_mean}
