% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{grades}
\alias{grades}
\title{Grades course Bioestatistical dataset}
\format{DataFrame}
\description{
Grades obtained from the course of General Biostatistics,courses from Universidad Nacional, Costa Rica.
}
\author{
Oscar Ramirez Alan \email{osoramirez@gmail.com}
}
\keyword{datasets}
