% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_mean.R
\name{h_mean}
\alias{h_mean}
\title{Harmonic mean}
\usage{
h_mean(x, na.rm = TRUE, zero = TRUE)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}

\item{na.rm}{is a na.rm}

\item{zero}{is any value of cero}
}
\value{
a geometric mean
}
\description{
get a Harmonic mean
}
\examples{
x<-rnorm(25,2,3)
h_mean(x)
}
\keyword{h_mean}
