% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K.R
\name{K}
\alias{K}
\title{A kurtosis}
\usage{
K(x, na.rm = FALSE)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector, matrix or data.frame}

\item{na.rm}{Should missing values be removed.}
}
\value{
A kurtosis
}
\description{
A skewness distribution,computes the estimator of Pearson's measure of kurtosis.
}
\examples{
x<-c(1,1,1,2,3,4,5,6,7,7,7,7,10,10)
K(x)
}
\keyword{kurtosis}
