% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
Mode
}
\description{
get a Mode
}
\examples{
x<-c(1,1,1,2,3,4,5,6,7,8,9,9,9,9,9)
Mode(x)
x2<-c(1,1,1,1,1,2,3,4,5,6,7,8,9,9,9,9,9)
Mode(x2)

}
\keyword{Mode}
