% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbox.R
\name{plotbox}
\alias{plotbox}
\title{plotbox}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
a plotbox
}
\description{
get a boxplot, showing the mean in plot
}
\examples{
plotbox(rnorm(14))
}
\keyword{plotbox}
