% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothist.R
\name{plothist}
\alias{plothist}
\title{histogram of data}
\usage{
plothist(x)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
a summary
}
\description{
Get a histogram, showing the mean and median of your data
}
\examples{
x<-rnorm(25,2,3)
plothist(x)


}
\keyword{plothist}
