% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{pvalue}
\usage{
pvalue()
}
\value{
pvalue
}
\description{
A pvalue.
}
\examples{
pvalue()
}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
#'@export pvalue
}
\keyword{pvalue}
