% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{resume}
\alias{resume}
\title{Summary Statitical Data}
\usage{
resume(x)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
a summary
}
\description{
Get a complete descriptive summary statistics
}
\examples{
set.seed(12345)
x<-rnorm(25,2,3)
resume(x)

}
\keyword{resume}
