% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume2data.R
\name{resume2data}
\alias{resume2data}
\title{Summary Statitical for your Data}
\usage{
resume2data(x)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}
}
\value{
a table summary
}
\description{
The most complete table descriptive in summary statistics
}
\examples{
set.seed(12345)
x<-rnorm(25,2,3)
resume2data(x)

}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
}
\keyword{resume2data}
