% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumendf.R
\name{resumendf}
\alias{resumendf}
\title{A Summary Statitical for Data.frame}
\usage{
resumendf(formula, data)
}
\arguments{
\item{formula}{ould be a  a vector, matrix or data.frame}

\item{data}{is your group}
}
\value{
a table summary
}
\description{
The table give a summary statistics
}
\examples{
data(iris)
resumendf(Petal.Width ~Species, data = iris)

}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
}
\keyword{resumendf}
