% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmout.R
\name{rmout}
\alias{rmout}
\title{Remove outlier}
\usage{
rmout(x, na.rm = TRUE)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}

\item{rmout}{Remove outlier}
}
\value{
a data set with oultiers
}
\description{
Remove outlier.
}
\examples{
set.seed(12345)
x <- rnorm(100)
x <- c(-15, x, 15)
y <- rmout(x)
par(mfrow = c(1, 2))
boxplot(x)
boxplot(y)
}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
}
