% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{A skewness}
\usage{
skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}

\item{na.rm}{Should missing values be removed.}
}
\value{
A skewness
}
\description{
A skewness distribution.
}
\examples{
x<-c(1,1,1,2,3,4,5,6,7,7,7,7,10,10)
skew(x)
}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
}
\keyword{skew}
