% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmout.R
\name{rmout}
\alias{rmout}
\title{Remove outlier}
\usage{
rmout(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{is a numeric value, could be a  a vector or data.frame}

\item{na.rm}{A logical scalar. Should missing values}

\item{...}{Additional arguments}
}
\value{
rmout
}
\description{
remove outlier
}
\examples{
set.seed(12345)
x <- rnorm(100)
x <- c(-15, x, 15)
y <- rmout(x)
}
\author{
Oscar Ramirez Alan (\email{osoramirez@gmail.com}).
}
